Moved KillZone.cpp and OneWayPlatform.cpp to the src folder
